refactor(phase-2): extract src/discord_ext/ and src/core/text|images
Split the two God-modules (src/utils.py, src/helpers.py) into focused
packages along a clean Discord / non-Discord boundary:

- src/core/text.py         — duration/number formatting, markdown escape,
                             sanitize, punctuation, search_dict_by_sentence,
                             extract_answer, pick_weighted_message
- src/core/images.py       — create_dynamic_image (Pillow)
- src/discord_ext/embeds.py       — Colors palette, SPACER_FIELD,
                                    format_discord_timestamp
- src/discord_ext/messages.py     — send_error/success, require_guild,
                                    fetch_user_safe, unarchive_if_thread,
                                    fetch_or_create_persistent_message
- src/discord_ext/autocomplete.py — guild_group_autocomplete,
                                    is_guild_enabled
- src/discord_ext/paginator.py    — CustomPaginator, format_poll,
                                    per-guild name_cache

src/utils.py and src/helpers.py are rewritten as re-export shims so
every existing "from src.utils import …" / "from src.helpers import …"
caller keeps working during the migration window.

Verified with a stubbed import probe (all 8 target modules plus both
shims import cleanly and re-export every expected symbol) and a full
compileall pass across src/ + extensions/ + main.py.

Refs #15 (Phase 2).
